feat: add unit degrees toggle
add radio buttons to the template.html form for
either fahrenheit or celsius selection.

modify apiCall() in index.js to take a new
parameter for the user's choice of unit type
(F or C) and inject into the fetch URL, and modify
the form submission listener to capture that
choice from the form before running apiCall().